[fix] --> error handling typo

+++ src/check_arg.py
@@ -35,8 +35,7 @@ def check_arg(abrs: list[str]) -> int:
             return 0
         sys.exit(84)
     if len(abrs) == 3:
-        if is_float(abrs[0] and is_int(abrs[1]) and is_int(abrs[2])) and int(abrs[0]) > 0 and int(
-                abrs[1]) > 0 and int(abrs[2]) > 0 and int(abrs[2]) > int(abrs[1]) > 0:
+        if is_float(abrs[0]) and is_int(abrs[1]) and is_int(abrs[2]) and float(abrs[0]) > 0 and int(abrs[1]) > 0 and int(abrs[2]) > 0 and int(abrs[2]) > int(abrs[1]) > 0:
             do_case2(float(abrs[0]), int(abrs[1]), int(abrs[2]))
             return 0
         return 84
